requirements.txt
Django==3.0.6
pywinrm==0.4.1
requests==2.23.0

psycopg2==2.8.5
psycopg2-binary==2.8.5

celery
redis
django-redis

pytest==5.4.2
pytest-django==3.9.0
pytest-cov==2.8.1
mixer==6.1.3
snapshottest==0.5.1
